+++ packages/core/src/index.ts
@@ -18,6 +18,7 @@ export {
   retryOnUploadNotComplete,
   getReleasesStatus,
   promoteRelease,
+  assignRelease,
   updateRollout,
   listTracks,
   createTrack,